requirements.txt
Django>=5.0.6,<5.1
psycopg2>=2.9.5,<2.10
Pillow>=10.4.0,<10.5
gunicorn>=20.1.0,<20.2
boto3>=1.34.144,<1.35
django-storages>=1.14.4,<1.15
dj-database-url>=2.2.0,<2.3
